renamed Home and Console containers to Landing and Home
