+++ src/containers/Home/Home.js
@@ -1,14 +1,49 @@
+import { useCallback, useEffect } from 'react';
+import { useHistory } from 'react-router-dom';
+import { Link } from 'react-router-dom';
+import { useSelector, useDispatch } from 'react-redux';
 import Hero from '../../components/Hero';
 import HeroTitle from '../../components/HeroTitle';
 import HeroSubtitle from '../../components/HeroSubtitle';
+import Gallery from '../../components/Gallery';
+import * as actions from '../../store/actions';
 
-const Home = () => {
+const Home = (props) => {
+  const history = useHistory();
+
+  const [auth, doodle] = useSelector((state) => [state.auth, state.doodle]);
+  const dispatch = useDispatch();
+
+  const fetchDoodles = useCallback(
+    () => dispatch(actions.doodlesRequestFetch(auth.id)),
+    [dispatch, auth],
+  );
+
+  const editDoodle = useCallback(
+    (id) => {
+      history.push(`/design/${id}`);
+    },
+    [history],
+  );
+
+  const deleteDoodle = useCallback(
+    (id) => dispatch(actions.doodlesRequestDelete(auth.id, id)),
+    [dispatch, auth],
+  );
+
+  useEffect(() => {
+    fetchDoodles();
+  }, [fetchDoodles]);
+
   return (
     <>
       <Hero>
-        <HeroTitle>Make Something Wonderful!</HeroTitle>
-        <HeroSubtitle>Create things with DoodleIt</HeroSubtitle>
+        <HeroTitle>Create without limits.</HeroTitle>
+        <HeroSubtitle>
+          <Link to="/design">Go to Designer</Link>
+        </HeroSubtitle>
       </Hero>
+      <Gallery items={doodle.doodles} edit={editDoodle} delete={deleteDoodle} />
     </>
   );
 };
