+++ src/containers/Landing/Landing.js
@@ -1,0 +1,16 @@
+import Hero from '../../components/Hero';
+import HeroTitle from '../../components/HeroTitle';
+import HeroSubtitle from '../../components/HeroSubtitle';
+
+const Landing = () => {
+  return (
+    <>
+      <Hero>
+        <HeroTitle>Make Something Wonderful!</HeroTitle>
+        <HeroSubtitle>Create things with DoodleIt</HeroSubtitle>
+      </Hero>
+    </>
+  );
+};
+
+export default Landing;
